mcp/weather_mcp_server/pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "python-dotenv>=1.1.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
